here comes the window (and not without the famous "random pixel fade" effect)

+++ CleanMe.sh
@@ -2,6 +2,6 @@
 
 rm -v -f head/over/heels/*.class
 
-rmdir head/over/heels
-rmdir head/over
-rmdir head
+[ -d head/over/heels ] && rmdir head/over/heels
+[ -d head/over ] && rmdir head/over
+[ -d head ] && rmdir head
